refactor: align adapter module boundaries
