+++ src/infra/adapters/sqlite/schema.rs
@@ -1,0 +1,1 @@
+pub(super) const MAIN_SCHEMA: &str = "main";
